removed react handling. need to fix i18n package not being installed, mnaybe simple make re will be enough
