+++ frontend/index.html
@@ -5,7 +5,7 @@
 		<title>Transcendence Frontend</title>
 	</head>
 	<body>
-		<div id="root"></div>
-		<script src="./src/main.tsx" type="module"></script>
+		<div id="app"></div>
+		<script src="./src/main.ts" type="module"></script>
 	</body>
 </html>
