+++ frontend/package.json
@@ -8,16 +8,10 @@
 		"serve": "vite preview --host 0.0.0.0"
 	},
 	"dependencies": {
-		"i18next": "^23.16.8",
-		"i18next-browser-languagedetector": "^7.2.2",
-		"react": "^18.0.0",
-		"react-dom": "^18.0.0",
-		"react-i18next": "^13.5.0"
+		"i18next": "^22.0.0",
+		"i18next-browser-languagedetector": "^7.0.0"
 	},
 	"devDependencies": {
-		"@types/babel__generator": "^7.27.0",
-		"@types/react": "^19.1.0",
-		"@vitejs/plugin-react": "^4.0.0",
 		"autoprefixer": "^10.0.0",
 		"postcss": "^8.0.0",
 		"tailwindcss": "^3.0.0",
