+++ frontend/src/i18n/i18n.ts
@@ -1,10 +1,8 @@
 import i18n from 'i18next';
-import { initReactI18next } from 'react-i18next';
 import LanguageDetector from 'i18next-browser-languagedetector';
 
 i18n
   .use(LanguageDetector)
-  .use(initReactI18next)
   .init({
     fallbackLng: 'en',
     debug: true,
