+++ frontend/tsconfig.json
@@ -2,7 +2,6 @@
   "compilerOptions": {
     "target": "ESNext",
     "module": "ESNext",
-    "jsx": "react-jsx",
     "strict": true,
     "moduleResolution": "Node",
     "esModuleInterop": true,
